+++ web-frontend/modules/database/components/view/grid/GridViewRow.vue
@@ -105,6 +105,7 @@
         @cell-mousedown-left="$emit('cell-mousedown-left', { row, field })"
         @cell-mouseover="$emit('cell-mouseover', { row, field })"
         @cell-mouseup-left="$emit('cell-mouseup-left', { row, field })"
+        @add-row-after="$emit('add-row-after', $event)"
       ></GridViewCell>
     </div>
   </RecursiveWrapper>