+++ changelog.md
@@ -14,6 +14,7 @@ For example:
 * OAuth 2 flows now support redirects to specific pages. [#1288](https://gitlab.com/bramw/baserow/-/issues/1288)
 * Add support for language selection and group invitation tokens for OAuth 2 and SAML. [#1293](https://gitlab.com/bramw/baserow/-/issues/1293)
 * Calendar / date field picker: Highlight the current date and weekend [#1128](https://gitlab.com/bramw/baserow/-/issues/1128)
+* Pressing shift + enter in a selected cell of the grid view creates a new row. [#1097](https://gitlab.com/bramw/baserow/-/issues/1097)
 
 ### Bug Fixes